WP-03: identifier engines with a verified safe default
Six engines: TCKN, VKN, IBAN, PAN, phone, email. Each exposes generate() and
is_checksum_valid(), and the second is the same function verify runs over
emitted files, so an engine cannot certify its own safety.

Two facts were verified before any code was written, not after.

The VKN check digit algorithm was taken from two independent open
implementations in different languages by different authors, cross-checked
against each other over 200 000 random inputs with zero disagreements, and
against the published vector 4540536920. A subtly wrong algorithm would still
emit ten plausible digits while silently breaking the safety claim, and a test
written against the same wrong algorithm would not notice.

The IBAN bank code 99999 was verified unassigned against the TCMB payment
systems participant list: 71 participants, four-digit codes spanning 0001 to
0807, nothing in the 9xxxx range. Under the validator policy the check digits
are deliberately correct, so the fictional bank code is the only thing keeping a
checksum-valid synthetic IBAN from naming a real institution.

Both records, with sources and retrieval dates, are in
docs/normative-verification.md, which also records a negative result: the
Turkish numbering plan reserves no fictional mobile range, so coincidence with
an assigned number cannot be engineered away and is documented instead.

Turkish folding is explicit rather than str.lower(), which expands the dotted
capital I into two code points and would shift every label span after it. A
property test asserts folding never changes a string's code point count.

+++ tools/gen_tables.py
@@ -100,7 +100,7 @@ def inverse_cdf_knots(spec: LogNormalSpec) -> list[int]:
         probability = (index + 0.5) / KNOTS
         # Inverse CDF of the log-normal: exp(mu + sigma * probit(p)).
         quantile = math.exp(spec.mu + spec.sigma * _probit(probability))
-        value = int(round(quantile))
+        value = round(quantile)
         knots.append(max(spec.minimum_kurus, min(spec.maximum_kurus, value)))
     return knots
